매크로 함수 오브젝트풀에도 적용

+++ LockFreeStack/CLockFreeStack.h
@@ -2,7 +2,7 @@
 #include "Windows.h"
 
 #define QueueSize 5000
-#define MAKE_NODE(value) ((unsigned long long)pReleaseNodeValue & (unsigned long long) 0x7FFFFFFFFFFF)
+#define MAKE_NODE(value) ((unsigned long long)pReleaseNodeValue & 0x7FFFFFFFFFFF)
 #define MAKE_VALUE(id, node) ((InterlockedIncrement(&_id) << 47) | (unsigned long long) pNewNode)
 
 enum EventType
